requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.6.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
pytz==2023.3
python-dotenv==1.0.0
alembic==1.12.1
email-validator==2.1.0.post1
httpx==0.25.1
jinja2==3.1.2
pyyaml==6.0.1
requests==2.31.0
boto3==1.29.0
pillow==10.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
pre-commit==3.3.3
httpx[http2]==0.25.1
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
fastapi-mail==1.9.0
python-magic==0.4.27
python-magic-bin==0.4.14
alembic==1.12.1
sqlalchemy==2.0.20
asyncpg==0.28.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
sentry-sdk==1.35.1
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-exporter-prometheus==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0